research/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aws_doc_ragqa"
version = "0.0.1"
description = "This repo contains all steps for deploying an application for Retrieval Augmented Generation Question Answering for AWS Documentation."
authors = [
  { name = "Leonardo Lara" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
requires-python = ">=3.11"
dependencies = [
    "boto3==1.39.4",
    "dotenv>=0.9.9",
    "llama-index-embeddings-bedrock==0.4.0",
    "llama-index-embeddings-gemini>=0.3.2",
    "llama-index-llms-anthropic>=0.6.19",
    "llama-index-llms-bedrock>=0.3.8",
    "llama-index-llms-gemini>=0.5.0",
    "llama-index-readers-file>=0.4.11",
    "llama-index-vector-stores-qdrant>=0.6.1",
    "nest-asyncio>=1.6.0",
    "weave>=0.51.56",
]


[tool.ruff]
line-length = 99
src = ["aws_doc_ragqa"]
include = ["pyproject.toml", "aws_doc_ragqa/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["aws_doc_ragqa"]
force-sort-within-sections = true

